Fixed refreshing issue
Fixed having to click home, was missing waiting for a promise to 
resolve. Login is now much faster as well as a result of not having to 
update the permissions. Permissions are loaded when the app is going to 
mount, then unsubcribing. We also resolve the user on login. Changed 
wording on login page as I was editing the same file at the same time.