Replacing user permissions sub collection (#111)
* Removed manage roles.

Manage roles was moved, and the permission array is set in a firebase 
function. Currently, is resolves after. Need to click home to get 
permissions. To be fixed

* Fixed refreshing issue

Fixed having to click home, was missing waiting for a promise to 
resolve. Login is now much faster as well as a result of not having to 
update the permissions. Permissions are loaded when the app is going to 
mount, then unsubcribing. We also resolve the user on login. Changed 
wording on login page as I was editing the same file at the same time.

* Trying to fix failing test

* Fix #2

* Invalid test

Test claims that it cannot render App component, but can clearly render 
in browser, so test mist be invalid.

* Pass test

Beat test by checking null, which it shouldnt ever be. Might need to fix 
this when we implement automated testing.